+++ src/cli/create.cpp
@@ -27,9 +27,9 @@ int create(Config *config) {
 
     while (!seq_reader->end_of_file()) {
         std::vector<common::SeqElem> read_seq_elems = seq_reader->get_aligned_seq_elements();
-        std::vector<std::pair<common::SeqElem, uint32_t>> seq_elems_to_insert;
+        std::vector<std::pair<common::SeqElem, uint64_t>> seq_elems_to_insert;
         for (const common::SeqElem &seq : read_seq_elems) {
-            seq_elems_to_insert.push_back(std::make_pair(seq, UINT_MAX));
+            seq_elems_to_insert.push_back(std::make_pair(seq, ULLONG_MAX));
         }
         ctc->insert_seq(seq_elems_to_insert);
     }
