+++ src/ds/static_types/static_acc_id.hpp
@@ -17,14 +17,14 @@ class AccessionIdSorted : public BaseSortedTreap {
 
     AccessionIdSorted(
         const std::string &req_key, 
-        const uint32_t req_database_id,
+        const uint64_t req_database_id,
         const int64_t req_seq_hash, 
         const int64_t req_metadata_hash
     );
     AccessionIdSorted(const H5::Group &h5_group);
     ~AccessionIdSorted();
 
-    static std::unique_ptr<BaseSortedTreap> get_unique_from_snapshot_line(const common::SeqElem &e, const uint32_t req_database_id, const BaseSortedTreap *);
+    static std::unique_ptr<BaseSortedTreap> get_unique_from_snapshot_line(const common::SeqElem &e, const uint64_t req_database_id, const BaseSortedTreap *);
     
     /*
      TODO FIX: because of this static function we have a problem when defining 2 different AccessionId treaps.
@@ -37,14 +37,14 @@ class AccessionIdSorted : public BaseSortedTreap {
                                                     const ds::PS_Treap *accid_snapshot_treap,
                                                     const ds::DB *base_db,
                                                     common::SeqElemReader*,
-                                                    const std::vector<uint32_t> &insertions_db_ids, 
-                                                    const std::vector<uint32_t> &deletions_db_ids, 
-                                                    std::vector<std::pair<uint32_t, uint32_t>> updates_db_ids);
+                                                    const std::vector<uint64_t> &insertions_db_ids, 
+                                                    const std::vector<uint64_t> &deletions_db_ids, 
+                                                    std::vector<std::pair<uint64_t, uint64_t>> updates_db_ids);
 
     static std::vector<uint64_t> seq_hash_database_ids;
     static std::vector<uint64_t> metadata_hash_database_ids;
-    static uint32_t next_hash_id;
-    static std::unique_ptr<BaseSortedTreap> get_new_BaseSortedTreap(const std::string &key, const uint32_t database_id);
+    static uint64_t next_hash_id;
+    static std::unique_ptr<BaseSortedTreap> get_new_BaseSortedTreap(const std::string &key, const uint64_t database_id);
     static void reset_get_new_BaseSortedTreap(const H5::Group &group);
 
     static std::unique_ptr<BaseSortedTreap> copy_specialized_static_field(const BaseSortedTreap*);
